refactor encoding module to support adding gpu
